requirements-dev.txt
pytest
pytest-flask
pytest-mock
mongomock
pytest-playwright
